Removed unecessary imports, cleaned up subclassing, removed unused dependencies
