+++ Sources/monitr/Codables.swift
@@ -1,7 +1,6 @@
 import PathKit
 import Cron
 import Foundation
-import Dispatch
 
 extension Path: Codable {
     public init(from decoder: Decoder) throws {
