+++ Sources/monitr/Monitrs/Monitr+Base.swift
@@ -9,26 +9,16 @@
 
 */
 
-import Foundation
-import PathKit
-import Cron
-import SwiftyBeaver
-import Dispatch
-
 /// Checks the downloads directory for new content to add to Plex
 class Monitr<M> where M: Media {
     /// The configuration to use for the monitor
     var config: Config
 
     var moveOperationQueue: MediaOperationQueue
-    var convertOperationQueue: MediaOperationQueue
-    var cronStart: CronJob!
-    var cronEnd: CronJob!
 
-    init(_ config: Config, moveOperationQueue: MediaOperationQueue, convertOperationQueue: MediaOperationQueue) throws {
+    init(_ config: Config, moveOperationQueue: MediaOperationQueue) throws {
         self.config = config
         self.moveOperationQueue = moveOperationQueue
-        self.convertOperationQueue = convertOperationQueue
 
         // Since this media is not convertible, let's just set this false and not deal with it
         self.config.convert = false
@@ -62,26 +52,5 @@ class Monitr<M> where M: Media {
         } else {
             config.logger.warning("Unknown operation type '\(type(of: operation))'")
         }
-    }
-
-    /**
-     Stop watching the downloads directory
-
-     - Parameter now: If true, kills any active media management. Defaults to false
-    */
-    public func shutdown() {
-        config.logger.info("Shutting down \(M.self) monitr.")
-        moveOperationQueue.isSuspended = true
-        convertOperationQueue.isSuspended = true
-        if moveOperationQueue.operationCount > 0 {
-            config.logger.info("Saving \(M.self) conversion queue")
-            try? moveOperationQueue.save(to: config.configFile.parent + "moveOperationQueue.json")
-        }
-        if convertOperationQueue.operationCount > 0 {
-            config.logger.info("Saving \(M.self) conversion queue")
-            try? convertOperationQueue.save(to: config.configFile.parent + "convertOperationQueue.json")
-        }
-        moveOperationQueue.cancelAllOperations()
-        convertOperationQueue.cancelAllOperations()
     }
 }
